+++ HelikopterGPC/ident.h
@@ -5,28 +5,28 @@
 typedef struct{
 ARX* model;
 int AB;
-double Beta;
-double Lambda;
-double T;
-double Alpha;
-double Sigma;
+float Beta;
+float Lambda;
+float T;
+float Alpha;
+float Sigma;
 
 Matrix* P;
 Matrix* KFiT;
 Matrix* KFiTP;
-double*Theta;
-double*Phi;
-double*VectK;
-double*memY;
-double*memU;
+float*Theta;
+float*Phi;
+float*VectK;
+float*memY;
+float*memU;
 
 }IdentObj;
 
 
 
-void IdentObj_Constructor(IdentObj* obj,ARX* model,double Beta,double Lambda);
+void IdentObj_Constructor(IdentObj* obj,ARX* model,float Beta,float Lambda);
 void IdentObj_Destructor(IdentObj* obj);
-void CalculateIdent(IdentObj *idobj, double u, double y);
+void CalculateIdent(IdentObj *idobj, float u, float y);
 
 #endif // IDENT
