+++ HelikopterGPC/lumatrixinv.h
@@ -2,8 +2,8 @@
 #define LUMATRIXINV
 #define   TINY 1.0e-20
 
-void inverse(double **mat, double **y, double *col, int *indx, int dim, double *vv);
-void ludcmp(double **a, int n, int *indx, double *d, double *vv);
-void lubksb(double **a, int n, int *indx, double *b);
+void inverse(float **mat, float **y, float *col, int *indx, int dim, float *vv);
+void ludcmp(float **a, int n, int *indx, float *d, float *vv);
+void lubksb(float **a, int n, int *indx, float *b);
 #endif // LUMATRIXINV
